importBookCsv

+++ src/main/java/com/project/LibManager/service/BookService.java
@@ -1,17 +1,29 @@
 package com.project.LibManager.service;
 
+import java.io.BufferedReader;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import org.apache.commons.csv.CSVFormat;
+import org.apache.commons.csv.CSVParser;
+import org.apache.commons.csv.CSVRecord;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.project.LibManager.dto.request.BookCreateRequest;
 import com.project.LibManager.dto.request.BorrowingRequest;
@@ -53,29 +65,37 @@ public class BookService {
     BorrowwingMapper borrowwingMapper;
 
     @Transactional
-    @PreAuthorize("hasRole('ADMIN')")
     public BookResponse createBook(BookCreateRequest bookCreateRequest) {
         if (bookCreateRequest == null) {
             log.error("BookCreateRequest is null");
-            throw new AppException(ErrorCode.INVALID_KEY);
+            throw new AppException(ErrorCode.UNCATEGORIZED_EXCEPTION);
         }
-        
+
         BookType type = bookTypeRepository.findById(bookCreateRequest.getTypeId())
-        .orElseThrow(() -> new AppException(ErrorCode.BOOKTYPE_NOT_EXISTED));
+                .orElseThrow(() -> new AppException(ErrorCode.BOOKTYPE_NOT_EXISTED));
 
         try {
-            Book book = bookMapper.toBook(bookCreateRequest);
-            book.setType(type);
-            bookRepository.save(book);   
-            return bookMapper.toBookResponse(book);
+            Optional<Book> existingBook = bookRepository.findByIsbn(bookCreateRequest.getIsbn());
+
+            if (existingBook.isPresent()) {
+                Book book = existingBook.get();
+                book.setStock(book.getStock() + bookCreateRequest.getStock());
+                bookRepository.save(book);
+                return bookMapper.toBookResponse(book);
+            } else {
+                Book book = bookMapper.toBook(bookCreateRequest);
+                book.setType(type);
+                bookRepository.save(book);
+                return bookMapper.toBookResponse(book);
+            }
         } catch (DataAccessException e) {
-            log.error("Database error: " + e.getMessage(), e);
+            log.error("Database error: {}", e.getMessage(), e);
             throw new AppException(ErrorCode.UNCATEGORIZED_EXCEPTION);
         }
     }
 
+
     @Transactional
-    @PreAuthorize("hasRole('ADMIN')")
     public BookResponse updateBook(BookCreateRequest bookCreateRequest, Long bookId) {
         if (bookCreateRequest == null) {
             log.error("BookCreateRequest is null");
@@ -83,16 +103,20 @@ public class BookService {
         }
 
         Book book = bookRepository.findById(bookId)
-        .orElseThrow(() -> new AppException(ErrorCode.BOOK_NOT_EXISTED));
+                .orElseThrow(() -> new AppException(ErrorCode.BOOK_NOT_EXISTED));
 
         BookType type = bookTypeRepository.findById(bookCreateRequest.getTypeId())
-        .orElseThrow(() -> new AppException(ErrorCode.BOOKTYPE_NOT_EXISTED));
-        
-        
+                .orElseThrow(() -> new AppException(ErrorCode.BOOKTYPE_NOT_EXISTED));
+
         try {
+            if (!book.getIsbn().equals(bookCreateRequest.getIsbn())) {
+                if (bookRepository.findByIsbn(bookCreateRequest.getIsbn()).isPresent()) {
+                    throw new AppException(ErrorCode.BOOK_EXISTED);
+                }
+            }
+
             bookMapper.updateBook(book, bookCreateRequest);
             book.setType(type);
-
             return bookMapper.toBookResponse(bookRepository.save(book));
         } catch (DataAccessException e) {
             log.error(e.getMessage());
@@ -100,12 +124,17 @@ public class BookService {
         }
     }
 
+
     @Transactional
     @PreAuthorize("hasRole('ADMIN')")
     public void deleteBook(Long id) {
         Book book = bookRepository.findById(id).orElseThrow(()->new AppException(ErrorCode.USER_NOT_EXISTED));
-        Set<Borrowing> setBorrow = book.getBorrowings();
-        setBorrow.clear();
+        
+        boolean isBorrowed = borrowingRepository.existsByBookAndReturnDateIsNull(book);
+        if (isBorrowed) {
+            throw new AppException(ErrorCode.BOOK_IS_CURRENTLY_BORROWED);
+        }
+
         try {
             bookRepository.delete(book);
         } catch (Exception e) {
@@ -241,5 +270,58 @@ public class BookService {
             throw new AppException(ErrorCode.UNCATEGORIZED_EXCEPTION);
         }
     }
-    
+
+    @Transactional
+    public void importBooks(MultipartFile file) {
+        if (file.isEmpty()) {
+            throw new RuntimeException("The file is empty. Please select a valid CSV file.");
+        }
+        if (file.getSize() > 5 * 1024 * 1024) {
+            throw new RuntimeException("The file is too large. Maximum allowed size is 5MB.");
+        }
+
+        try (BufferedReader reader = new BufferedReader(new InputStreamReader(file.getInputStream(), StandardCharsets.UTF_8))) {
+            CSVParser csvParser = CSVFormat.DEFAULT
+                    .builder()
+                    .setHeader()
+                    .setSkipHeaderRecord(true)
+                    .setIgnoreSurroundingSpaces(true)
+                    .build()
+                    .parse(reader);
+
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy/MM/d");
+
+            Map<String, Book> booksToUpdate = new HashMap<>();
+            List<Book> newBooks = csvParser.getRecords().stream().map(record -> {
+                String isbn = record.get("isbn");
+                return bookRepository.findByIsbn(isbn)
+                        .map(book -> {
+                            book.setStock(book.getStock() + Integer.parseInt(record.get("stock")));
+                            booksToUpdate.put(isbn, book);
+                            return book;
+                        })
+                        .orElseGet(() -> Book.builder()
+                                .isbn(isbn)
+                                .title(record.get("title"))
+                                .author(record.get("author"))
+                                .type(bookTypeRepository.findById(Long.parseLong(record.get("typeId")))
+                                        .orElseThrow(() -> new AppException(ErrorCode.BOOKTYPE_NOT_EXISTED)))
+                                .stock(Integer.parseInt(record.get("stock")))
+                                .publisher(record.get("publisher"))
+                                .publishedDate(LocalDate.parse(record.get("publishedDate"), formatter))
+                                .maxBorrowDays(Integer.parseInt(record.get("maxBorrowDays")))
+                                .location(record.get("location"))
+                                .coverImageUrl(record.get("coverImageUrl"))
+                                .build());
+            }).collect(Collectors.toList());
+            
+
+            bookRepository.saveAll(booksToUpdate.values());
+            bookRepository.saveAll(newBooks);
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+            throw new AppException(ErrorCode.UNCATEGORIZED_EXCEPTION);
+        }
+    }
+
 }
